+++ packages/halo/.storybook/config.js
@@ -1,3 +1,5 @@
+import React from 'react'
+import { css } from 'glamor'
 import requireContext from 'require-context.macro'
 
 import { addDecorator, configure } from '@storybook/react'
@@ -5,6 +7,14 @@ import { addDecorator, configure } from '@storybook/react'
 import centerDecorator from '@pluralsight/ps-design-system-storybook-addon-center'
 import themeDecorator from '@pluralsight/ps-design-system-storybook-addon-theme'
 
+const ConstrainWidth = props => (
+  <div
+    {...css({ label: 'constrain-width-decorator', maxWidth: 500 })}
+    {...props}
+  />
+)
+
+addDecorator(storyFn => <ConstrainWidth>{storyFn()}</ConstrainWidth>)
 addDecorator(centerDecorator)
 addDecorator(themeDecorator)