+++ packages/errors/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.21](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-errors@7.0.20...@pluralsight/ps-design-system-errors@7.0.21) (2020-01-23)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-errors
+
+
+
+
+
+## [7.0.20](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-errors@7.0.18...@pluralsight/ps-design-system-errors@7.0.20) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-errors
+
+
+
+
+
+## [7.0.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-errors@7.0.16...@pluralsight/ps-design-system-errors@7.0.18) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-errors
+
+
+
+
+
+## [7.0.16](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-errors@7.0.14...@pluralsight/ps-design-system-errors@7.0.16) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-errors
+
+
+
+
+
+## [7.0.14](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-errors@7.0.13...@pluralsight/ps-design-system-errors@7.0.14) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-errors
+
+
+
+
+
 ## [7.0.13](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-errors@7.0.12...@pluralsight/ps-design-system-errors@7.0.13) (2020-01-17)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-errors