+++ packages/note/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.24](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-note@5.0.23...@pluralsight/ps-design-system-note@5.0.24) (2020-01-23)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-note
+
+
+
+
+
+## [5.0.23](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-note@5.0.22...@pluralsight/ps-design-system-note@5.0.23) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-note
+
+
+
+
+
+## [5.0.22](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-note@5.0.20...@pluralsight/ps-design-system-note@5.0.22) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-note
+
+
+
+
+
+## [5.0.20](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-note@5.0.18...@pluralsight/ps-design-system-note@5.0.20) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-note
+
+
+
+
+
+## [5.0.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-note@5.0.16...@pluralsight/ps-design-system-note@5.0.18) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-note
+
+
+
+
+
+## [5.0.16](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-note@5.0.15...@pluralsight/ps-design-system-note@5.0.16) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-note
+
+
+
+
+
 ## [5.0.15](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-note@5.0.14...@pluralsight/ps-design-system-note@5.0.15) (2020-01-17)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-note