+++ packages/tab/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [14.0.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tab@14.0.17...@pluralsight/ps-design-system-tab@14.0.18) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tab
+
+
+
+
+
+## [14.0.17](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tab@14.0.15...@pluralsight/ps-design-system-tab@14.0.17) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tab
+
+
+
+
+
+## [14.0.15](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tab@14.0.13...@pluralsight/ps-design-system-tab@14.0.15) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tab
+
+
+
+
+
+## [14.0.13](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tab@14.0.11...@pluralsight/ps-design-system-tab@14.0.13) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tab
+
+
+
+
+
+## [14.0.11](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tab@14.0.10...@pluralsight/ps-design-system-tab@14.0.11) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tab
+
+
+
+
+
 ## [14.0.10](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tab@14.0.9...@pluralsight/ps-design-system-tab@14.0.10) (2020-01-17)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-tab