+++ packages/typeahead/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.17](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-typeahead@3.2.16...@pluralsight/ps-design-system-typeahead@3.2.17) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-typeahead
+
+
+
+
+
+## [3.2.16](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-typeahead@3.2.14...@pluralsight/ps-design-system-typeahead@3.2.16) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-typeahead
+
+
+
+
+
+## [3.2.14](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-typeahead@3.2.12...@pluralsight/ps-design-system-typeahead@3.2.14) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-typeahead
+
+
+
+
+
+## [3.2.12](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-typeahead@3.2.10...@pluralsight/ps-design-system-typeahead@3.2.12) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-typeahead
+
+
+
+
+
+## [3.2.10](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-typeahead@3.2.9...@pluralsight/ps-design-system-typeahead@3.2.10) (2020-01-22)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-typeahead
+
+
+
+
+
 ## [3.2.9](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-typeahead@3.2.8...@pluralsight/ps-design-system-typeahead@3.2.9) (2020-01-17)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-typeahead